# Update algorithm to process 8 bytes at a time