--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -2,14 +2,14 @@ name: CI
 
 on:
   push:
-    branches: main
-    tags: v*
+    branches: [main]
+    tags: ["v*"]
   pull_request:
-    branches: "*"
+    branches: ["*"]
 
 jobs:
-  all:
-    name: all
+  push:
+    name: push
     runs-on: ubuntu-latest
 
     steps: