This algorithm is adapted from: https://github.com/komrad36/CRC#option-9-8-byte-tabular

The 8-byte variant is chosen somewhat arbitrarily as the default implementation rather than the 16-byte variant, because the memory requirement is less and both variants have reasonable performance. (Unclear why there is a small difference in the main exported implementation and the other 8-byte implementation in benchmarks. :shrug:)

Adds tests to ensure the results are the same as the naïve implementation regardless of data length and alignment.

Results from my M1 MacBook Air (16GB RAM):

```
$ node --version
v15.14.0
$ yarn bench
yarn run v1.22.10
$ ts-node --project tsconfig.cjs.json bench/index.ts
Validating CRC implementations:
┌─────────┬────────────┬────────────┬────────────┬────────────┬────────────┬────────────┬────────────┐
│ (index) │  current   │   naive    │   1 byte   │   2 byte   │   4 byte   │   8 byte   │  16 byte   │
├─────────┼────────────┼────────────┼────────────┼────────────┼────────────┼────────────┼────────────┤
│    0    │ 1545002771 │ 1545002771 │ 1545002771 │ 1545002771 │ 1545002771 │ 1545002771 │ 1545002771 │
└─────────┴────────────┴────────────┴────────────┴────────────┴────────────┴────────────┴────────────┘
Results match.
Running "Table generation" suite...
Progress: 100%

  ternary:
    387 704 ops/s, ±0.77%   | slowest, 51.59% slower

  branchless:
    565 049 ops/s, ±0.95%   | 29.44% slower

  branchless unrolled:
    800 850 ops/s, ±0.57%   | fastest

Finished 3 cases!
  Fastest: branchless unrolled
  Slowest: ternary
Running "CRC" suite...
Progress: 100%

  naive:
    16 ops/s, ±0.31%      | slowest, 99.38% slower

  1 byte:
    352 ops/s, ±0.27%     | 86.34% slower

  2 bytes:
    560 ops/s, ±0.31%     | 78.26% slower

  4 bytes:
    1 053 ops/s, ±0.43%   | 59.12% slower

  8 bytes:
    1 785 ops/s, ±7.34%   | 30.71% slower

  16 bytes:
    2 576 ops/s, ±0.36%   | fastest

  current implementation:
    2 017 ops/s, ±0.43%   | 21.7% slower

  current implementation, unaligned:
    2 016 ops/s, ±0.40%   | 21.74% slower

Finished 8 cases!
  Fastest: 16 bytes
  Slowest: naive
✨  Done in 67.45s.
```